New markdown-it plugin to show marked spans
To show search results with matching substrings highlighted, the
markdown renderer needs to be supplied a variable set of terms to
highlight rather than being coded in the markdown.

Interface for the plugin comes in two sections, the function that
markdown-it uses to modify its renderer, and some helpers that will
create patterns to be passed as the environment second argument to
render each call.

Released under Apache-2.0 license.